refactor default game stuff and game mods table
